requirements.txt
beautifulsoup4==4.6.3
certifi==2018.11.29
chardet==3.0.4
Django==2.1.3
django-js-asset==1.1.0
django-mptt==0.9.1
django-taggit==0.23.0
geoip2==2.9.0
idna==2.7
lxml==4.2.5
maxminddb==1.4.1
mysqlclient==1.3.10
Pillow==5.3.0
python-decouple==3.1
pytz==2018.7
requests==2.20.1
sorl-thumbnail==12.5.0
urllib3==1.24.1
